+++ packages/api/src/platforms/vtex/clients/commerce/types/Region.ts
@@ -1,16 +1,22 @@
 export interface RegionInput {
-  postalCode: string;
-  country: string;
-  salesChannel?: string | null;
+  postalCode: string
+  geoCoordinates: GeoCoordinates | null
+  country: string
+  salesChannel?: string | null
 }
 
 export interface Seller {
-  id: string; // storeframework01
-  name: string; // My Awsome Seller
-  logo: string;
+  id: string // storeframework01
+  name: string // My Awsome Seller
+  logo: string
+}
+
+export interface GeoCoordinates {
+  latitude: GLfloat
+  longitude: GLfloat
 }
 
 export interface Region {
-  id: string;
-  sellers: Seller[];
+  id: string
+  sellers: Seller[]
 }