+++ packages/components/src/organisms/Filter/FilterSlider.tsx
@@ -1,5 +1,5 @@
-import type { HTMLAttributes } from 'react'
-import React, { PropsWithChildren } from 'react'
+import type { HTMLAttributes, PropsWithChildren } from 'react'
+import React from 'react'
 
 import {
   Button,
@@ -35,6 +35,10 @@ export interface FilterSliderProps extends HTMLAttributes<HTMLDivElement> {
    */
   clearBtnProps?: Partial<ButtonProps>
   /**
+   * Props forwarded to the `Overlay` component.
+   */
+  overlayProps?: OverlayProps
+  /**
    * Function called when Close Button is clicked.
    */
   onClose: () => void
@@ -51,8 +55,8 @@ function FilterSlider({
   children,
   applyBtnProps,
   clearBtnProps,
-  onClose,
   overlayProps,
+  onClose,
   ...otherProps
 }: PropsWithChildren<FilterSliderProps>) {
   const { fade, fadeOut } = useFadeEffect()