+++ packages/components/src/organisms/SlideOver/SlideOver.tsx
@@ -30,6 +30,10 @@ export interface SlideOverProps extends Omit<ModalProps, 'title'> {
   fade: 'in' | 'out'
   children: ReactNode
   /**
+   * Props forwarded to the `Overlay` component.
+   */
+  overlayProps?: OverlayProps
+  /**
    * This function is called whenever the user clicks outside.
    * the modal content
    */
@@ -41,14 +45,14 @@ export interface SlideOverProps extends Omit<ModalProps, 'title'> {
 }
 
 function SlideOver({
+  testId = 'fs-slide-over',
   isOpen,
   direction = 'leftSide',
   size = 'full',
   fade = 'out',
   children,
-  onDismiss,
-  testId = 'fs-slide-over',
   overlayProps,
+  onDismiss,
   ...otherProps
 }: SlideOverProps) {
   return (